Extract DbSchemaExport class from DbExport
